+++ app/Http/Controllers/API/PhysicalSwitchController.php
@@ -58,6 +58,8 @@ class PhysicalSwitchController extends Controller
 
     public function massDestroy(MassDestroyPhysicalSwitchRequest $request)
     {
+        abort_if(Gate::denies('physical_switch_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         PhysicalSwitch::whereIn('id', request('ids'))->delete();
 
         return response(null, Response::HTTP_NO_CONTENT);
