+++ app/Http/Controllers/API/TaskController.php
@@ -58,6 +58,8 @@ class TaskController extends Controller
 
     public function massDestroy(MassDestroyTaskRequest $request)
     {
+        abort_if(Gate::denies('task_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         Task::whereIn('id', request('ids'))->delete();
 
         return response(null, Response::HTTP_NO_CONTENT);
